requirements.txt
influxdb-client==1.49.0
pandas==2.3.1
opencv-python==4.12.0.88
timeout-decorator==0.5.0
pyserial==3.5
cobs==1.2.1
Sphinx==8.2.3
gpiozero==2.0.1
smbus2==0.5.0
crcmod==1.7
lgpio==0.2.2.0
watchdog-6.0.0
ultralytics-8.4.21
